requirements.txt
# Core dependencies
telethon==1.36.0
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Optional but recommended
# httpx  # For FastAPI TestClient (if needed for testing)



